+++ source/calculators/weighted.cpp
@@ -63,6 +63,8 @@ EstOutput Weighted::createProcesses(Tree* t, vector< vector<string> > namesOfGro
 	try {
         vector<int> processIDS;
 		EstOutput results;
+        bool recalc = false;
+        
 #if defined (__APPLE__) || (__MACH__) || (linux) || (__linux) || (__linux__) || (__unix__) || (__unix)
 		int process = 1;
 		
@@ -88,12 +90,82 @@ EstOutput Weighted::createProcesses(Tree* t, vector< vector<string> > namesOfGro
 				
 				exit(0);
 			}else { 
-				m->mothurOut("[ERROR]: unable to spawn the necessary processes."); m->mothurOutEndLine(); 
-				for (int i = 0; i < processIDS.size(); i++) { kill (processIDS[i], SIGINT); }
-				exit(0);
+                m->mothurOut("[ERROR]: unable to spawn the number of processes you requested, reducing number to " + toString(process) + "\n"); processors = process;
+                for (int i = 0; i < processIDS.size(); i++) { kill (processIDS[i], SIGINT); }
+                //wait to die
+                for (int i=0;i<processIDS.size();i++) {
+                    int temp = processIDS[i];
+                    wait(&temp);
+                }
+                m->control_pressed = false;
+                for (int i=0;i<processIDS.size();i++) {
+                    m->mothurRemove(outputDir + (toString(processIDS[i]) + ".weightedcommand.results.temp"));
+                }
+                recalc = true;
+                break;
 			}
 		}
 	
+        if (recalc) {
+            //test line, also set recalc to true.
+            //for (int i = 0; i < processIDS.size(); i++) { kill (processIDS[i], SIGINT); } for (int i=0;i<processIDS.size();i++) { int temp = processIDS[i]; wait(&temp); } m->control_pressed = false;  for (int i=0;i<processIDS.size();i++) {m->mothurRemove(outputDir + (toString(processIDS[i]) + ".weightedcommand.results.temp"));}processors=3; m->mothurOut("[ERROR]: unable to spawn the number of processes you requested, reducing number to " + toString(processors) + "\n");
+            
+            lines.clear();
+            
+            //calculate number of comparisons i.e. with groups A,B,C = AB, AC, BC = 3;
+            int numGroups = m->getNumGroups();
+            vector< vector<string> > namesOfGroupCombos;
+            for (int i=0; i<numGroups; i++) {
+                for (int l = 0; l < i; l++) {
+                    //initialize weighted scores
+                    vector<string> groups; groups.push_back((m->getGroups())[i]); groups.push_back((m->getGroups())[l]);
+                    namesOfGroupCombos.push_back(groups);
+                }
+            }
+            
+            int remainingPairs = namesOfGroupCombos.size();
+            int startIndex = 0;
+            for (int remainingProcessors = processors; remainingProcessors > 0; remainingProcessors--) {
+                int numPairs = remainingPairs; //case for last processor
+                if (remainingProcessors != 1) { numPairs = ceil(remainingPairs / remainingProcessors); }
+                lines.push_back(linePair(startIndex, numPairs)); //startIndex, numPairs
+                startIndex = startIndex + numPairs;
+                remainingPairs = remainingPairs - numPairs;
+            }
+            
+            results.clear();
+            processIDS.resize(0);
+            process = 1;
+            
+            //loop through and create all the processes you want
+            while (process != processors) {
+                pid_t pid = fork();
+                
+                if (pid > 0) {
+                    processIDS.push_back(pid);  //create map from line number to pid so you can append files in correct order later
+                    process++;
+                }else if (pid == 0){
+                    EstOutput Myresults;
+                    Myresults = driver(t, namesOfGroupCombos, lines[process].start, lines[process].num, ct);
+                    
+                    //pass numSeqs to parent
+                    ofstream out;
+                    string tempFile = outputDir + m->mothurGetpid(process) + ".weighted.results.temp";
+                    m->openOutputFile(tempFile, out);
+                    
+                    out << Myresults.size() << endl;
+                    for (int i = 0; i < Myresults.size(); i++) {  out << Myresults[i] << '\t';  } out << endl;
+                    out.close();
+                    
+                    exit(0);
+                }else { 
+                    m->mothurOut("[ERROR]: unable to spawn the necessary processes."); m->mothurOutEndLine(); 
+                    for (int i = 0; i < processIDS.size(); i++) { kill (processIDS[i], SIGINT); }
+                    exit(0);
+                }
+            }
+        }
+        
 		results = driver(t, namesOfGroupCombos, lines[0].start, lines[0].num, ct);
 	
 		//force parent to wait until all the processes are done
